requirements.txt
openai>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
tqdm>=4.62.0
python-dotenv>=0.19.0
chromadb==1.0.8
fastapi>=0.101.0
uvicorn[standard]>=0.23.0
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
tavily-python>=0.2.5
protobuf>=4.25.0
pydantic>=2.0.0
plotly
requests>=2.31.0
